Init rewrite commit for version 0.3.0
Versions 0.2.x will be 'beta' or development. Following the even/odd
versioning so far followed.

Going to migrate/edit/update most of the code.
Also going to change the design to allow for argument parsing

Check the notes.txt file to see a TODO list on what should be
accomplished